changed the gtm back to index
